% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{assert_logical}
\alias{assert_logical}
\title{Assert a value is a single logical}
\usage{
assert_logical(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to test.}

\item{name}{Name to use in the error message.}
}
\description{
Assert a value is a single logical
}
\keyword{internal}
