% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{assert_non_negative_number}
\alias{assert_non_negative_number}
\title{Assert a value is a single non-negative number}
\usage{
assert_non_negative_number(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to test.}

\item{name}{Name to use in the error message.}
}
\description{
Assert a value is a single non-negative number
}
\keyword{internal}
