% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{assert_positive_integer}
\alias{assert_positive_integer}
\title{Assert a value is a single positive integer}
\usage{
assert_positive_integer(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to test.}

\item{name}{Name to use in the error message.}
}
\description{
Assert a value is a single positive integer
}
\keyword{internal}
