% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{assert_service_params}
\alias{assert_service_params}
\title{Assert service_params has valid structure for all 3 urgency levels}
\usage{
assert_service_params(service_params)
}
\arguments{
\item{service_params}{List with keys "1", "2", "3". Each element must be a
list with positive numeric \code{mean} and \code{sd}.}
}
\description{
Assert service_params has valid structure for all 3 urgency levels
}
\keyword{internal}
