% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{assert_string}
\alias{assert_string}
\title{Assert a value is a non-empty character string}
\usage{
assert_string(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to test.}

\item{name}{Name to use in the error message.}
}
\description{
Assert a value is a non-empty character string
}
\keyword{internal}
