% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_interarrival}
\alias{draw_interarrival}
\title{Draw an inter-arrival time from an exponential distribution}
\usage{
draw_interarrival(rate)
}
\arguments{
\item{rate}{Arrival rate (patients per time unit).}
}
\value{
Non-negative numeric inter-arrival time.
}
\description{
Draw an inter-arrival time from an exponential distribution
}
\keyword{internal}
