% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ERsim-package}
\alias{ERsim}
\alias{ERsim-package}
\title{ERsim: Emergency Room Queue Simulation}
\description{
Discrete event simulation tools for modelling patient flow through an
emergency room.
}
\author{
\strong{Maintainer}: David Abraham \email{david@example.com} (487868)

Authors:
\itemize{
  \item David Abraham \email{david@example.com} (487868)
  \item Mevin Biju (486097)
  \item Akhil Ambalapurath Ajith (487806)
}

}
