% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_patient_id}
\alias{make_patient_id}
\title{Generate a unique patient ID string}
\usage{
make_patient_id(n)
}
\arguments{
\item{n}{Sequential patient number.}
}
\value{
Character ID, e.g. "P00042".
}
\description{
Generate a unique patient ID string
}
\keyword{internal}
