% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{make_resources}
\alias{make_resources}
\title{Build a list of Resource objects for a given role}
\usage{
make_resources(n, role = c("doctor", "nurse"))
}
\arguments{
\item{n}{Integer. Number of resources to create.}

\item{role}{Character. "doctor" or "nurse".}
}
\value{
List of \code{Resource} R6 objects.
}
\description{
Build a list of Resource objects for a given role
}
\keyword{internal}
