% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.SimConfig}
\alias{print.SimConfig}
\title{Print a SimConfig object}
\usage{
\method{print}{SimConfig}(x, ...)
}
\arguments{
\item{x}{A \code{SimConfig} object.}

\item{...}{Ignored.}
}
\description{
Print a SimConfig object
}
