% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print.SimResults}
\alias{print.SimResults}
\title{Print a SimResults Object}
\usage{
\method{print}{SimResults}(x, ...)
}
\arguments{
\item{x}{A \code{SimResults} object.}

\item{...}{Ignored.}
}
\description{
Print a SimResults Object
}
