% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{summary.SimResults}
\alias{summary.SimResults}
\title{Summary of a SimResults Object}
\usage{
\method{summary}{SimResults}(object, ...)
}
\arguments{
\item{object}{A \code{SimResults} object.}

\item{...}{Ignored.}
}
\description{
Summary of a SimResults Object
}
