% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{urgency_label}
\alias{urgency_label}
\title{Convert urgency integer to label}
\usage{
urgency_label(urgency)
}
\arguments{
\item{urgency}{Integer urgency level (1, 2, or 3).}
}
\value{
Character label.
}
\description{
Convert urgency integer to label
}
\keyword{internal}
