% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_sim_config}
\alias{validate_sim_config}
\title{Validate a SimConfig object}
\usage{
validate_sim_config(config)
}
\arguments{
\item{config}{A \code{SimConfig} object created by \code{new_sim_config}.}
}
\value{
The config object invisibly, or an informative error.
}
\description{
Called automatically by \code{\link{new_sim_config}}. Can also be called
directly to re-validate a config that has been manually modified.
}
