% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{assert_positive_number}
\alias{assert_positive_number}
\title{Assert a value is a single positive number}
\usage{
assert_positive_number(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to test.}

\item{name}{Name to use in the error message.}
}
\description{
Assert a value is a single positive number
}
\keyword{internal}
